requirements.txt
Flask
gunicorn
azure-cosmos
pyodbc
python-dotenv
requests
reportlab
pillow